+++ common/src/plugins/resources/axis/axis_resource.cpp
@@ -103,7 +103,7 @@ bool QnPlAxisResource::readMotionInfo()
     {
         QMutexLocker lock(&m_mutex);
         qint64 time = qnSyncTime->currentMSecsSinceEpoch();
-        if (time - m_lastMotionReadTime < MOTION_INFO_UPDATE_INTERVAL)
+        if (time - m_lastMotionReadTime < qint64(MOTION_INFO_UPDATE_INTERVAL))
             return true;
 
         m_lastMotionReadTime = time;
